✨ Refactor member request and schema files; consolidate private member functions and enhance type definitions for improved clarity and functionality
