--- packages/client/src/request/member.ts
@@ -1,6 +1,0 @@
-import { useFetch } from '@/hooks/use-fetch';
-import { PublicMemberListResponse } from '@/schemas/member';
-
-export const useMemberList = () => {
-  useFetch<PublicMemberListResponse>('/members');
-};
